+++ src/cross_hs.h
@@ -46,9 +46,9 @@ class HS : public QTLCross
     const std::vector<std::string> geno_names(const std::vector<std::string> alleles,
                                               const bool is_x_chr);
 
-    const double nrec(const int gen_left, const int gen_right,
-                      const bool is_x_chr, const bool is_female,
-                      const Rcpp::IntegerVector& cross_info);
+    const int nrec(const int gen_left, const int gen_right,
+                   const bool is_x_chr, const bool is_female,
+                   const Rcpp::IntegerVector& cross_info);
 };
 
 #endif // CROSS_HS_H
